pkg/test,pkg/util: move e2eutil from util to test
