+++ test/test-framework/memcached_test.go
@@ -20,9 +20,9 @@ import (
 	"testing"
 	"time"
 
-	operator "github.com/operator-framework/operator-sdk/test/test-framework/pkg/apis/cache/v1alpha1"
 	framework "github.com/operator-framework/operator-sdk/pkg/test"
-	"github.com/operator-framework/operator-sdk/pkg/util/e2eutil"
+	"github.com/operator-framework/operator-sdk/pkg/test/e2eutil"
+	operator "github.com/operator-framework/operator-sdk/test/test-framework/pkg/apis/cache/v1alpha1"
 
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/types"
